Parameterize button count and make undermount a joystick property
The joystick module needs to know the undermount value so the dust cover
ring gets placed correctly. Therefore, pass undermount right through to
the joystick and don't do the adjustment anywhere else.

Also, put the button positions into an array so that a different button
count can be selected for each control panel.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>