Move control panel box into separate module
The control panel itself is logically separate from a box used to encase
it. Some cabinet designs will forgo the box entirely. Others will
consist of only the box. Split the box into a separate file entirely to
show the difference.

As part of spliting the box, also refactor the code generating the
control panel. Parameterize the multilayer panel generation so that it
is simpler to use, and extrude panels from a 2D panel profile shape.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>