Remove the globals for control panel layout
The control panel module used a bunch of global values that made it
difficult to change the size of a panel design. Remove the globals and
start passing the values around as parameters instead. There is still
some refactoring needed to get rid of the control placement angle.

The multilayer panel code also got refactored to eliminate the explicit
depth of each layer. Instead the function is called recursively to stack
the panels on top of each other at 0.2mm separation.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>