Add more control layout options and show examples
Add support for multiple layout options for each control cluster and
change the demo code to show a sample of each control option when
viewing controls.scad in OpenSCAD. Each layout is given a name that is
printed with the layout in the sample view.

At the same time rename button_pad() to control_cluster() to reflect the
new functionality.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>