Replace 'cutout' argument with 'action' on controls
'cutout' was a boolean that indicated if the control is being added to the
view, or if the control cutout outlines were needed for carving the control
panel. However, a boolean is limiting because it doesn't allow for any other
operations to be added to the module. Follow the lead of Jeremie Francois'
excellent OpenSCAD tutorial[1] and rename the argument to 'action', with 'add'
meaning the control itself and 'remove' meaning the negative space cutout. Also
add 'guide' for creating a drilling guide for holes.

[1] http://www.tridimake.com/2014/11/how-to-use-openscad-4-children-and.html

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>