Fix control box to use size & inset arguments correctly
The control_box module hasn't been updated since the control_panel was
switched to using 'size' and 'inset' arguments instead of width, depth,
cab_width, and inset. Therefore the box being generated didn't match the
dimensions of the panel() module when created. Fix the code to at least
get the right dimensions. It still needs to be fixed use the right
outline profile, but at least the size is correct now.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>