Fix placement of control clusters on curved panel
The cluster placement code didn't correctly calculate the position of
clusters on a curve, and wasn't very good at spacing to get maximum
panel utilization. Fix the placement code to correctly calculate the arc
length and angle between clusters for a curved board, and to use a
defined value to place clusters as close to the sides as possible.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>