Convert panel() module to use an action argument
The panel() module was only taking a 'show_controls' bool argument to
control the output, but there are several kinds of output needed,
including a cut plane for angled panels, dimension lines, and the
aforementioned controls. Convert the module to accept an action argument
and restructure the code to suit.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>