Reimplement cutlines() to be faster and more accurate
cutlines() was just a slow function, making use of minkowski() to create
a thin outline drawing of the panel. By using projection() to get a 2D
slice of the panel, the 2D offset() operation can be used to get a
simple cut profile that lies outside the shape border.

Outside lines are arguably better because cutting tools have a
thickness, and they need to follow a path outside of the actual shape to
cut the right size.  It could also be argued that this is irrelevant for
the cutlines() module because it produces double-walled lines that are
never going to be used for cutting. Rather, DXF output is created
directly from getting a projection() of the panel shape.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>