Carve control cutouts individually for each layer
To prepare for more accurate generation of lasercut lines, change the
panel_multilayer() function to carve out the control mounting holes on
each layer individually instead of all at once. The next step will be to
add a mode to panel_multilayer that calculates and outputs the 2D
lasercut curves instead of 3D objects.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>