Refactor panel depth calculations
Refactor panel_multilayer to use a for() loop instead of recursive
calls. This simplifies the code a bit. Also change the way layers are
calculated. Subtract a margin of layer_gap height from the top and
bottom of every layer so that control placement doesn't have to try and
guess where margins need to be added. The controls can work on the layer
heights as given, and the panel will subtract a margin so that rendering
works correctly.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>