Add manufacturing output for the 4 player box
MFG files used on 8 Aug 2017 to build the panel for a 4 player box. A
few sizes need to be modified after cutting this design. The trackball
hole was slightly too small, as was the joystick cutout. Those two
parameters will be adjusted before the next time this panel is cut.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>