Make all dimensions 2D objects
Dimensions are 2D constructs, and are being output in an SVG file which
can only handle 2D objects. It doesn't make sense for them to be
generated as 3D objects when they immediately need to be projected back
into 2D. Make all of dimlines create 2D objects so that generating
output is easier and faster.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>